Resolve conflicts with upstream main:

- torch_custom_ops.py: drop the removed
  `tensorrt_llm.plugin.plugin.CustomAllReduceHelper` import (main now
  imports it from `_torch.distributed.allreduce_helper`) and keep the
  quantization.mode validation helpers import.

- linear.py: keep both sides' new module-level helpers. Adopt main's
  `W4A8NVFP4FP8LinearMethod.load_weight_scales` signature (`module`,
  `shard_keys`) and carry `quant_mode` through it, updating the three
  call sites accordingly.

- create_moe.py: main relocated the MoE package from
  `_torch/modules/fused_moe` to `_torch/moe/fused_moe`, so the FP4
  architecture check moves to the new `create_moe`. The old
  `can_implement` guard is dropped: backend eligibility is now owned by
  `resolve_moe_cls`, which raises with the rejection trail when
  `allow_backend_degradation` is False.

+++ tests/unittest/_torch/thop/parallel/test_fp8_rowwise_linear.py
@@ -42,7 +42,7 @@ def test_fp8_rowwise_linear(dtype):
     with torch.inference_mode():
         x_dq = x_fp8.to(x_scale.dtype) * x_scale.view(-1, 1)
         w_dq = w_fp8.to(w_scale.dtype).t() * w_scale.view(1, -1)
-        ref_output = x_dq.to(dtype) @ w_dq.to(dtype)
+        ref_output = (x_dq.float() @ w_dq.float()).to(dtype)
 
     # compare
     torch.cuda.synchronize()
